this is the first build
let m know about any changes you think I should implement . the end of the game and the current start was just due to figuring out a way for the collection of coins to be the condition for the end of the game. Bbasically all of the more time consuming parts of the game have been completed so anything ontop of this should be done easily before 2 or 3 for another upload of the build. i will be working on the build from here on , trying to improve the overall level as (like I said) it is the most basic form of what was needed.